+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -83,6 +83,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added fade factor for decal projectors.
 - Added stereo instancing macros to most shaders used in VR
 - Added multi edition support for HDRenderPipelineAsset
+- Added emissive property for shader graph decals
 
 ### Fixed
 - Fixed logic to disable FPTL with stereo rendering
@@ -154,6 +155,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed pre-exposure incorrectly apply to preview
 - Fixed issue with duplicate 3D texture in 3D texture altas of volumetric
 - Fixed Camera rendering order (base on the depth parameter)
+- Fixed shader graph decals not being cropped by gizmo
 - Fixed "Attempting to draw with missing UAV bindings" errors on Vulkan.
 
 ### Changed