+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added emissive property for decal material (non-shader graph)
 - Added cascade visualizer, supporting disabled handles when not overriding.
 - Added support of motion vector for transparent object (as replace)
+- Added emissive property for shader graph decals
 
 ### Fixed
 - Fixed a few UI bugs with the color grading curves.
@@ -114,7 +115,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added a Lighting and BackLighting slots in Lit, StackLit, Fabric and Hair master nodes
 - Added xrInstancing flag support to RTHandle
 - Added support for cullmask for decal projectors
-- Added emissive property for shader graph decals
 
 ### Fixed
 - Fixed logic to disable FPTL with stereo rendering